+++ src/common/bu/EventBuilder.cc
@@ -33,7 +33,7 @@ doProcessing_(false)
 
 void evb::bu::EventBuilder::addSuperFragment
 (
-  const uint32_t buResourceId,
+  const uint16_t buResourceId,
   FragmentChainPtr& superFragments
 )
 {
